+++ CHANGELOG.md
@@ -9,9 +9,13 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### New
 
 ### Changes
+- removed modules as they now are in `https://github.com/awslabs/seedfarmer-modules`
 
 ### Fixes
-
+- fix for `gitpython` assuming that codecommit repos are unsafe 
+- fix resolution of parameters that have a False value
+- fix adding verbose error messages when a manifest cannot be loaded
+- fix destroy of modules sourced from git via disparate (non-cached) cli invocations
 
 ## v2.6.0 (2023-03-10)